writing_improver/requirements.txt
flask
pyyaml
python-dotenv
openai
nltk
tqdm
